Move the player AI into its own class, keeping TurnHandling focused on turn handling mechanics.
